requirements.txt
google-cloud-vision==3.4.4
google-api-python-client==2.100.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.1.0
pandas==2.1.1
openpyxl==3.1.2
python-dotenv==1.0.0
Pillow==10.0.1
